#57 moved delete logic to delete acooutn button in settings
